refactor(server-action): rename common folder
