+++ src/app/quiz/create/page.tsx
@@ -1,6 +1,6 @@
 import { ErrorModal } from "../../../components/modal/ErrorModal";
 import { QuizForm } from "../../../components/quiz/QuizForm";
-import { getSignInInfo } from "../../../server-action/backend/common/cookie";
+import { getSignInInfo } from "../../../server-action/backend/_common/cookie";
 
 export default async function QuizCreatePage() {
 	const user = await getSignInInfo();
