+++ src/components/auth/DeleteAccountForm.tsx
@@ -3,8 +3,8 @@
 import { useState, FormEvent } from "react";
 import { deleteUserAction } from "../../server-action/backend/user/api";
 import toast from "react-hot-toast";
-import { isHttpException, isServerActionError } from "../../server-action/backend/common/util";
-import { HTTP_STATUS } from "../../server-action/backend/common/status";
+import { isHttpException, isServerActionError } from "../../server-action/backend/_common/util";
+import { HTTP_STATUS } from "../../server-action/backend/_common/status";
 import { GuideModal } from "../modal/GuideModal";
 import { useRouter } from "next/navigation";
 import { ErrorModal } from "../modal/ErrorModal";
