+++ src/components/auth/PasswordUpdateForm.tsx
@@ -2,9 +2,9 @@
 
 import { useState } from "react";
 import { useRouter } from "next/navigation";
-import { HTTP_STATUS } from "../../server-action/backend/common/status";
-import { HttpException, ServerActionError } from "../../server-action/backend/common/dto";
-import { isHttpException, isServerActionError } from "../../server-action/backend/common/util";
+import { HTTP_STATUS } from "../../server-action/backend/_common/status";
+import { HttpException, ServerActionError } from "../../server-action/backend/_common/dto";
+import { isHttpException, isServerActionError } from "../../server-action/backend/_common/util";
 import toast from "react-hot-toast";
 
 export interface PasswordUpdateFormProps {
