+++ src/components/auth/SignUpForm.tsx
@@ -4,8 +4,8 @@ import { useState } from "react";
 import toast from "react-hot-toast";
 import { signUpAction } from "../../server-action/backend/user/api";
 
-import { isHttpException, isServerActionError } from "../../server-action/backend/common/util";
-import { HTTP_STATUS } from "../../server-action/backend/common/status";
+import { isHttpException, isServerActionError } from "../../server-action/backend/_common/util";
+import { HTTP_STATUS } from "../../server-action/backend/_common/status";
 import { useRouter } from "next/navigation";
 import { GuideModal } from "../modal/GuideModal";
 import { ErrorModal } from "../modal/ErrorModal";
