+++ src/server-action/backend/artist/api.ts
@@ -1,7 +1,7 @@
 "use server";
 import { CondOperator, RequestQueryBuilder } from "@dataui/crud-request";
-import { HttpException, IPaginationResult } from "../common/dto";
-import { getServerUrl, withErrorHandler } from "../common/lib";
+import { HttpException, IPaginationResult } from "../_common/dto";
+import { getServerUrl, withErrorHandler } from "../_common/lib";
 import { Artist } from "./dto";
 
 const getArtists = async (
